+++ CHANGELOG.md
@@ -7,6 +7,14 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations
 on how to structure this file.
 This plugin uses [Semantic versioning](https://semver.org).
 
+## [2.11.1] - 2022-01-13
+
+### Fixed
+
+- Solved issue that the query link was set on the wrong object
+- Solved some issues that the plugin crashed when working on OMT documents 
+  (incomplete imports for example)
+
 ## [2.11.0] - 2022-01-13
 
 ### Added